Conflicts:
	app/controllers/gp_article/admin/docs_controller.rb

+++ app/models/gp_article/doc.rb
@@ -309,6 +309,7 @@ class GpArticle::Doc < ActiveRecord::Base
   end
 
   def close_page(options={})
+    return true if will_replace?
     return false unless super
     publishers.destroy_all unless publishers.empty?
     FileUtils.rm_rf(::File.dirname(public_path))
@@ -374,6 +375,7 @@ class GpArticle::Doc < ActiveRecord::Base
     when :replace
       new_doc.prev_edition = self
       new_doc.in_tasks = self.in_tasks
+      new_doc.in_creator = {'group_id' => creator.group_id, 'user_id' => creator.user_id}
     else
       new_doc.name = nil
       new_doc.title = new_doc.title.gsub(/^(【複製】)*/, '【複製】')
@@ -461,7 +463,11 @@ class GpArticle::Doc < ActiveRecord::Base
   def editable?
     result = super
     return result unless result.nil? # See "Sys::Model::Auth::EditableGroup"
-    return editable_group.all?
+    return editable_group.all? || approval_participators.include?(Core.user)
+  end
+
+  def publishable?
+    super || approval_participators.include?(Core.user)
   end
 
   def formated_display_published_at
